+++ gnome/extensions/.local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/lib/Manager.js
@@ -2,7 +2,7 @@
  * Manager Library
  *
  * @author     Javad Rahmatzadeh <j.rahmatzadeh@gmail.com>
- * @copyright  2020-2024
+ * @copyright  2020-2025
  * @license    GPL-3.0-only
  */
 
@@ -113,6 +113,14 @@ export class Manager
 
         this.#settings.connect('changed::quick-settings-dark-mode', () => {
             this.#applyQuickSettingsDarkMode(false);
+        });
+
+        this.#settings.connect('changed::quick-settings-night-light', () => {
+            this.#applyQuickSettingsNightLight(false);
+        });
+
+        this.#settings.connect('changed::quick-settings-airplane-mode', () => {
+            this.#applyQuickSettingsAirplaneMode(false);
         });
 
         this.#settings.connect('changed::window-picker-icon', () => {
@@ -302,6 +310,18 @@ export class Manager
         this.#settings.connect('changed::max-displayed-search-results', () => {
             this.#applyMaxDisplayedSearchResults(false);
         });
+
+        this.#settings.connect('changed::accent-color-icon', () => {
+            this.#applyAccentColorIcon(false);
+        });
+
+        this.#settings.connect('changed::workspace-thumbnail-to-main-view', () => {
+            this.#applyWorkspaceThumbnailToMainView(false);
+        });
+
+        this.#settings.connect('changed::invert-calendar-column-items', () => {
+            this.#applyInvertCalendarColumnItems(false);
+        });
     }
 
     /**
@@ -325,6 +345,8 @@ export class Manager
         this.#applyAccessibilityMenu(false);
         this.#applyQuickSettings(false);
         this.#applyQuickSettingsDarkMode(false);
+        this.#applyQuickSettingsNightLight(false);
+        this.#applyQuickSettingsAirplaneMode(false);
         this.#applyWindowPickerIcon(false);
         this.#applyTypeToSearch(false);
         this.#applyWorkspaceSwitcherSize(false);
@@ -369,6 +391,9 @@ export class Manager
         this.#applyWorkspacePeek(false);
         this.#applyDashAppRunning(false);
         this.#applyMaxDisplayedSearchResults(false);
+        this.#applyAccentColorIcon(false);
+        this.#applyWorkspaceThumbnailToMainView(false);
+        this.#applyInvertCalendarColumnItems(false);
     }
 
     /**
@@ -392,6 +417,8 @@ export class Manager
         this.#applyAccessibilityMenu(true);
         this.#applyQuickSettings(true);
         this.#applyQuickSettingsDarkMode(true);
+        this.#applyQuickSettingsNightLight(true);
+        this.#applyQuickSettingsAirplaneMode(true);
         this.#applyWindowPickerIcon(true);
         this.#applyTypeToSearch(true);
         this.#applyWorkspaceSwitcherSize(true);
@@ -436,6 +463,9 @@ export class Manager
         this.#applyWorkspacePeek(true);
         this.#applyDashAppRunning(true);
         this.#applyMaxDisplayedSearchResults(true);
+        this.#applyAccentColorIcon(true);
+        this.#applyWorkspaceThumbnailToMainView(true);
+        this.#applyInvertCalendarColumnItems(true);
     }
 
     /**
@@ -681,6 +711,38 @@ export class Manager
             this.#api.quickSettingsDarkStyleToggleShow();
         } else {
             this.#api.quickSettingsDarkStyleToggleHide();
+        }
+    }
+
+    /**
+     * apply quick settings night light
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    #applyQuickSettingsNightLight(forceOriginal)
+    {
+        if (forceOriginal || this.#settings.get_boolean('quick-settings-night-light')) {
+            this.#api.quickSettingsNightLightToggleShow();
+        } else {
+            this.#api.quickSettingsNightLightToggleHide();
+        }
+    }
+
+    /**
+     * apply quick settings airplane mode
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    #applyQuickSettingsAirplaneMode(forceOriginal)
+    {
+        if (forceOriginal || this.#settings.get_boolean('quick-settings-airplane-mode')) {
+            this.#api.quickSettingsAirplaneModeToggleShow();
+        } else {
+            this.#api.quickSettingsAirplaneModeToggleHide();
         }
     }
 
@@ -1475,6 +1537,54 @@ export class Manager
             this.#api.setMaxDisplayedSearchResultToDefault();
         } else {
             this.#api.setMaxDisplayedSearchResult(items);
+        }
+    }
+
+    /**
+     * apply accent color icon
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    #applyAccentColorIcon(forceOriginal)
+    {
+        if (forceOriginal || !this.#settings.get_boolean('accent-color-icon')) {
+            this.#api.accentColorIconDisable();
+        } else {
+            this.#api.accentColorIconEnable();
+        }
+    }
+
+    /**
+     * apply workspace thumbnail to main view
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    #applyWorkspaceThumbnailToMainView(forceOriginal)
+    {
+        if (forceOriginal || !this.#settings.get_boolean('workspace-thumbnail-to-main-view')) {
+            this.#api.workspaceThumbnailClickToDefault();
+        } else {
+            this.#api.workspaceThumbnailClickToMainView();
+        }
+    }
+
+    /**
+     * apply invert calendar column items
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    #applyInvertCalendarColumnItems(forceOriginal)
+    {
+        if (forceOriginal || !this.#settings.get_boolean('invert-calendar-column-items')) {
+            this.#api.revertCalendarColumnItemsToDefault();
+        } else {
+            this.#api.invertCalendarColumnItems();
         }
     }
 }