+++ gnome/extensions/.local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/CHANGELOG.md
@@ -7,6 +7,50 @@ We go to the next version after each release on the
 
 ## [Unreleased]
 
+## [34.0] - 2025-02-02
+
+### Removed
+
+- The monthly notification support.
+
+## [33.0] - 2025-02-02
+
+### Fixed
+
+- The monthly notification support calculation.
+
+## [32.0] - 2025-02-01
+
+### Added
+
+- Airplane Mode Toggle Button Visibility in Quick Settings.
+- GNOME Shell 48 Support.
+- More Crypto Donations (XRP, Solana, Cardano, BNB, Stellar, USDC, LTC, Shiba Inu and LBRY).
+- Support Notifier.
+
+### Fixed
+
+- Looking Glass `_resize()` issue.
+
+## [31.0] - 2025-01-06
+
+### Added
+
+- Accent Color for Icons.
+- Crypto Support to the Preferences Window.
+- Invert Calendar Column Items.
+- Night Light Toggle Button Visibility in Quick Settings.
+- Workspace Switcher Click Always Goes to The Main View.
+- Ukrainian Translation by [@vovkiv](https://gitlab.gnome.org/vovkiv).
+
+### Fixed
+
+- Search margin when using high display scale factor and panel in overview feature.
+- Slowest animation.
+- The dash app running indicator visibility (GNOME Shell 46 and 47).
+- The Looking Glass Position Using Inspector in Overview.
+- The Workspaces Shadow in App Grid When the Search is Disabled.
+
 ## [30.0] - 2024-07-31
 
 ### Added